existente SubRule gelöscht; ausgelagert in neue SubRule, die noch komplett unfertig ist
git-svn-id: svn://localhost/trunk/c++/mathematics/cas2@271 99334987-7fb6-4f9b-9453-cf9f8c1b9a5c

+++ SimplifyTerm.cpp
@@ -34,7 +34,7 @@ void CAS::SimplifyWithRule(CAS::TermReference* ref, CAS::RuleCollection* rule, s
       ++index;
       iterators[index] = childrenSimplifications[index].begin();
     }
-    for (int i = 0; i < paramcount; ++i)
+    for (size_t i = 0; i < paramcount; ++i)
       dataArray[i] = (*iterators[i])->Clone();
     TermReference *tTerm = new TermReference (ref->get_const()->CreateTerm (dataArray));
     if (!coll.push_back (tTerm))
